+++ src/index.js
@@ -1,66 +1,37 @@
 import fs from "fs";
 import path from "path";
 import { v4 as uuidv4 } from "uuid";
-import { COMPRESSION_LEVEL, zip } from "zip-a-folder";
 
-import configJson from "../config/config.json" assert { type: "json" };
 import ConfigValidator from "./ConfigValidatorClass.js";
 import DocumentFactory from "./DocumentFactoryClass.js";
-
-function getTodayDate() {
-  const today = new Date();
-  const year = today.getFullYear();
-  let month = today.getMonth() + 1;
-  let day = today.getDate();
-
-  if (day < 10) day = "0" + day;
-  if (month < 10) month = "0" + month;
-
-  const formattedToday = year + "-" + month + "-" + day;
-
-  return formattedToday;
-}
-
-function writeDocument(outputDir, jsonDocument) {
-  // if folder doesn't exist, create one
-  if (!fs.existsSync(outputDir)) {
-    fs.mkdirSync(outputDir, { recursive: true });
-  }
-
-  const uniqueFileId = uuidv4();
-
-  const outputPath = path.join(outputDir, `${Date.now()}-${uniqueFileId}.json`);
-  const stringifiedDocument = JSON.stringify(jsonDocument);
-  fs.writeFileSync(outputPath, stringifiedDocument, "utf8");
-}
-
-async function zipFolder(srcDir, destDir) {
-  await zip(srcDir, destDir, { compression: COMPRESSION_LEVEL.high });
-}
+import { getTodayDate } from "./utils/dateUtils.js";
+import { writeDocumentToDir, zipFolder } from "./utils/fileUtils.js";
 
 async function init() {
   const uniqueFolderId = uuidv4();
-  const { schemaPath, nullablePercentage, documentCount, outputDir, references } = configJson;
-  const schemaFile = JSON.parse(fs.readFileSync(schemaPath));
-  const { schema, derivatives } = schemaFile;
-  const targetFolder = path.join(outputDir, getTodayDate(), uniqueFolderId);
+  const configJson = JSON.parse(fs.readFileSync("./config/config.json"));
+  const { recipeFilePath, nullablePercentage, documentCount, outputDir, references } = configJson;
 
-  console.info(`Successfully retrieved schema from: '${schemaPath}'. Performing validation...`);
+  const recipeFile = JSON.parse(fs.readFileSync(recipeFilePath));
+  const { schema, derivatives } = recipeFile;
+  const destinationFolder = path.join(outputDir, getTodayDate(), uniqueFolderId);
+
+  console.info(`Successfully retrieved recipe from: '${recipeFilePath}'. Performing validation...`);
   console.info("-----------------------------------------------------------------------------");
   console.info("Errors found:");
   const validator = new ConfigValidator(schema, derivatives, references);
   validator.validateSchema();
   validator.validateDerivatives();
-  const isValidSchema = validator.getValidity();
+  const isValidRecipe = validator.getValidity();
   console.info("-----------------------------------------------------------------------------");
 
-  if (!isValidSchema) {
-    console.error("Something is wrong with the provided schema, exiting program...");
+  if (!isValidRecipe) {
+    console.error("Something is wrong with the provided recipe, exiting program...");
     return;
   }
 
   console.info(
-    `Validation successful! Generating ${documentCount} documents to ${targetFolder}.zip`,
+    `Validation successful! Generating ${documentCount} documents to ${destinationFolder}`,
   );
 
   for (let i = 0; i < documentCount; i++) {
@@ -70,12 +41,12 @@ async function init() {
       references,
       derivatives,
     ).getDocument();
-    writeDocument(targetFolder, newDocument);
+    writeDocumentToDir(destinationFolder, newDocument);
   }
 
   // zip and delete folder afterwards
-  await zipFolder(targetFolder, `${targetFolder}.zip`);
-  fs.rmSync(targetFolder, { recursive: true, force: true });
+  await zipFolder(destinationFolder, `${destinationFolder}`);
+  fs.rmSync(destinationFolder, { recursive: true, force: true });
 
   console.info(`Successfully generated ${documentCount} documents!`);
 }