+++ src/utils/dateUtils.js
@@ -1,0 +1,13 @@
+export function getTodayDate() {
+  const today = new Date();
+  const year = today.getFullYear();
+  let month = today.getMonth() + 1;
+  let day = today.getDate();
+
+  if (day < 10) day = "0" + day;
+  if (month < 10) month = "0" + month;
+
+  const formattedToday = year + "-" + month + "-" + day;
+
+  return formattedToday;
+}